refactor: remove duplicated validateProposalMessage function
